fix Anthropic answer parsing move embedding selector to general util
